refactor propcard into a standalone component file
